[BUGFIX] Fixed the terminal box dimensions
The previous dimensions did not work with this font, it would fallback to 1040x768. After experimenting I found that the height needed to be 774px or higher (AND the width 1046px or higher which it already was) for the terminal box to render with the provided dimensions and not the fallback dimensions. No clue why that extra 6px is required for both the dimensions.